fix(critical): ALL reads on daily_subscriptions fail — amount int64 vs numeric(12,2)
ROOT CAUSE (found via live API test):
  curl GET /api/v1/admin/daily-subscriptions returned:
    'sql: Scan error on column index 4, name "amount":
     converting driver.Value type string ("20.00") to a int64: invalid syntax'

  DailySubscription entity had:
    Amount int64 `gorm:"column:amount;type:numeric(12,2)"`

  PostgreSQL returns numeric(12,2) values as strings (e.g. "20.00").
  Go's database/sql driver cannot scan "20.00" into int64 → scan error.
  This error propagates from EVERY query to the daily_subscriptions table:
    - getUserSubscriptions  → returns []  → dashboard shows 0 subscriptions
    - getSubscriptionStats  → returns 0   → header card shows 0
    - GetAllSubscriptions   → fails       → admin panel errors
    - FindActiveByMSISDN    → fails       → subscription status endpoint fails
    - FindByPaymentRef      → fails       → ProcessFirstPayment fails
                                           → subscription stays 'pending' forever

  This single type mismatch was the root cause of ALL subscription issues.

FIXES:
  1. entities/daily_subscriptions.go: Amount int64 → float64
     float64 correctly scans PostgreSQL numeric(n,m) values.

  2. subscription_service.go: Amount creation value uses float64(kobo)/100
     Fallback amt uses int64(sub.Amount * 100) to convert naira → kobo.

  3. user_service.go: SubscriptionItem.Amount int64 → float64
     getSubscriptionStats cast: int64(sub.Amount) for the summary total.

+++ backend/internal/application/services/user_service.go
@@ -109,7 +109,7 @@ type SubscriptionItem struct {
 	ID              uuid.UUID `json:"id"`
 	TransactionDate time.Time `json:"transaction_date"`
 	Reference       string    `json:"reference"`
-	Amount          int64     `json:"amount"`
+	Amount          float64   `json:"amount"` // naira value from numeric(12,2) column
 	Entries         int       `json:"entries"`
 	PointsEarned    int       `json:"points_earned"`
 	Status          string    `json:"status"`
@@ -830,7 +830,7 @@ func (s *UserService) getSubscriptionStats(ctx context.Context, msisdn string) (
 	for _, sub := range subs {
 		count++
 		if sub.Status == "active" {
-			amount += sub.Amount
+			amount += int64(sub.Amount) // Amount is naira float64; cast to int64 for summary
 			// entries and points from lifetime totals on the sub row
 			entries += int64(sub.TotalEntries)
 			points += int64(sub.TotalPointsAwarded)
